+++ my-app/src/helpers/constants/actions.ts
@@ -1,0 +1,9 @@
+export enum ActionKind {
+  ADD_NAMES = 'ADD_NAMES',
+  ADD_CHARACTERS = 'ADD_CHARACTERS',
+  INIT_LOADING = 'INIT_LOADING',
+  LOAD_CHARACTERS_STATE = 'LOAD_CHARACTERS_STATE',
+  CHANGE_PAGE = 'CHANGE_PAGE',
+  CHANGE_SEARCH_VALUE = 'CHANGE_SEARCH_VALUE',
+  ADD_CARD = 'ADD_CARD',
+}
